fix(fuzz): update fuzz targets for u128 supply accumulators (#333)
The #333 u128 widening of the chain-state supply accumulators
(total_mined, total_fees_burned) missed the workspace-excluded fuzz/
crate, breaking the Fuzz Build Guard (PR):

- fuzz_add_block.rs: `prev.total_mined + expected_reward` no longer
  compiles (u128 + u64); cast the u64 reward to u128. Drop the now
  redundant `as u128` casts on total_mined / total_fees_burned (both
  already u128) so clippy stays green. lottery_pool stays u64, so its
  cast is preserved.
- fuzz_cluster_tax_math.rs: calculate_block_reward's total_supply param
  is now u128; cast the fuzzer's u64 supply at the call site.

No accumulator is narrowed with `as u64`. Verified `cargo fuzz build`
compiles every target clean on the pinned nightly-2025-12-03 toolchain,
and the six #333 regression tests still pass.

+++ fuzz/fuzz_targets/fuzz_add_block.rs
@@ -104,7 +104,7 @@ fuzz_target!(|data: &[u8]| {
             // Invariant 3b: total_mined grew by exactly the block reward.
             assert_eq!(
                 post.total_mined,
-                prev.total_mined + expected_reward,
+                prev.total_mined + expected_reward as u128,
                 "accepted block did not increase total_mined by reward \
                  (prev={}, post={}, reward={})",
                 prev.total_mined,
@@ -121,9 +121,9 @@ fuzz_target!(|data: &[u8]| {
                 .get_lottery_pool()
                 .expect("lottery pool readable after successful add_block");
             let utxo_sum = sum_utxo_values(&ledger);
-            let accounted = utxo_sum + post.total_fees_burned as u128 + lottery_pool as u128;
+            let accounted = utxo_sum + post.total_fees_burned + lottery_pool as u128;
             assert_eq!(
-                post.total_mined as u128, accounted,
+                post.total_mined, accounted,
                 "supply conservation violated after accepted block: \
                  total_mined={} != utxo_sum={} + burned={} + pool={}",
                 post.total_mined, utxo_sum, post.total_fees_burned, lottery_pool
